remove pointless doc type impls
